+++ src/clj/wil/routes/home.clj
@@ -5,7 +5,8 @@
   [hato.client :as hc]
   [ring.util.http-response :as response]
   [wil.layout :as layout]
-  [wil.middleware :as middleware]))
+  [wil.middleware :as middleware]
+  [wil.notes :refer [list-notes]]))
 
 (defn home-page
   [request]
@@ -19,10 +20,13 @@
   "retrieve str login's info from API.
    note: parameter is a string. cf. (db/get-user {:login login})"
   [login]
-  (let [ep   (str api-user login)
-        body (:body (hc/get ep {:as :json}))]
+  (let [body (:body (hc/get (str api-user login) {:as :json}))]
     (log/info "api-user" body)
     body))
+
+(comment
+  (get-user "hkimura")
+  )
 
 (defn login-page [request]
   (layout/render request "login.html" {:flash (:flash request)}))
@@ -53,11 +57,19 @@
     (-> (response/found "/login")
         (assoc :flash "please login"))))
 
+;; (defn list-texts
+;;   "admin 専用メソッド。引数 date の wil 一覧"
+;;   [{{:keys [date]} :path-params}]
+;;   (let [body (:body (hc/get (str "/api/list/" date) {:as :json}))]
+;;     (for [b body]
+;;       (:text b))))
+
 (defn home-routes []
   [""
    {:middleware [middleware/wrap-csrf
                  middleware/wrap-formats]}
-   ["/"       {:get home-page}]
-   ["/login"  {:get  login-page :post login-post}]
-   ["/logout" {:get logout}]
-   ["/profile" {:get profile-page}]])
+   ["/"        {:get home-page}]
+   ["/login"   {:get  login-page :post login-post}]
+   ["/logout"  {:get logout}]
+   ["/profile" {:get profile-page}]
+   ["/list/:date" {:get list-notes}]])